Update
* Increase coverage of backup tests
* FormatDSN
* Rewrite C.in to types for Stringer implementation
* Add PRAGMA func
* Fix TestPinger
* Driver, DriverContext => Fix Extensions, ConnectHook
* Add Stringer to CryptEncoder, CryptSaltedEncoder
* Fix OpenConnector
* Renamed context to vtable_context and include build tag
