+++ driver/driver.goconvey
@@ -1,0 +1,4 @@
+
+-tags=sqlite_userauth
+
+-cover
